agent/requirements.txt
# LiveKit agent dependencies for real-time AI video tutor
livekit-agents[deepgram,openai,cartesia,silero]~=1.4.4
pytest>=7.0.0
pytest-asyncio>=0.21.0
livekit-plugins-deepgram>=1.4.4
livekit-plugins-openai>=1.4.4
livekit-plugins-cartesia>=1.4.4
livekit-plugins-silero>=1.4.4
livekit-plugins-simli>=1.4.4
python-dotenv>=1.0.0
